Ajout des 8 planètes dans le main avec leur duréeRévolutionJours et changement des types float en double pour duréeRévolutionJours car sinon ça ne fonctionne pas

+++ src/main/java/HorlogeUniv/Astres/Astres.java
@@ -2,13 +2,13 @@ package HorlogeUniv.Astres;
 
 
 public abstract class Astres {
-    private float dureeRevolutionJours;
+    private double dureeRevolutionJours;
 
-    public Astres(float dureeRevolutionJours) {
+    public Astres(double dureeRevolutionJours) {
         this.dureeRevolutionJours = dureeRevolutionJours;
     }
 
-    public float getDureeRevolutionJours() {
+    public double getDureeRevolutionJours() {
         return dureeRevolutionJours;
     }
